Change id from int to long

+++ grails-app/services/com/ratchethealth/patient/EmailService.groovy
@@ -52,7 +52,7 @@ class EmailService extends RatchetAPIService {
         }
     }
 
-    def emailCheck(String token, int id, int last4Number) {
+    def emailCheck(String token, long id, int last4Number) {
         String checkPhoneNumberUrl = grailsApplication.config.ratchetv2.server.url.email.checkPhoneNumber
 
         def url = String.format(checkPhoneNumberUrl, id)
@@ -73,7 +73,7 @@ class EmailService extends RatchetAPIService {
         }
     }
 
-    def subscribe(String token, int id, int last4Number, boolean subscribe) {
+    def subscribe(String token, long id, int last4Number, boolean subscribe) {
         String subscribeUrl = grailsApplication.config.ratchetv2.server.url.email.subscribe
 
         def url = String.format(subscribeUrl, id)
